Add beginnings of build_pe_installer task
This commit adds a Bolt task that will build a pe-installer package to
be used in PE. At the moment, it can build a package based on the
current state of pe-installer-vanagon, or it can be pointed to local
changes for both the enterprise_tasks and higgs repositories. Ideally,
this should also work when pointed to PRs for those repos.

This commit also contains some testing code, like moving the package to
my desktop when the task is done...