Add pe-modules support, fix requires/directory structure
This commit adds a task for building custom pe-modules packages, with
valid parameters being any repository that pe-modules-vanagon has as a
component. These can be pointed to changes in both local and PR forms.

This commit also edits some of the requires for the build_package helper
functions, and moves those helper functions into a `files/` directory in
order to be more in line with the Bolt docs.