Add git branching logic
This commit adds the important logic for actually switching to the
correct branch of some arbitrary repo, based on the parameters that were
given and the repo that was checked out. Previously, this code would
just try to merge a PR into whatever the default branch of the repo is.
Obviously, that's not always ideal. This commit should theoretically fix
that, although it's absolutely not been tested thoroughly