Add tasks for provisioning a floaty host and pulling down a PE tarball
This commit adds a task to provision an arbitrary number of VMPooler
VMs, although at the moment it probably can't actually handle every type
of VM that vmpooler has. If a user passes in the exact right string,
it'll be fine, but when this task is part of a plan, there'll probably
need to be some manual verification to make sure pe_repo platform names
don't error out with getting floaty VMs and vice-versa. That'll be fun.

This commit also adds a task to pull down a PE tarball, it can handle
users passing in a codename, a PE family, a release version or a
specific rc- version, which I think should handle basically all the
valid cases for PE build types.