Fix version_tmp assignment and curl call
The curl call to pull down a PE build wouldn't fail if given a path to a
build that didn't exist, so this commit updates the curl command to
account for that. It also fixes a bug where it was possible for
version_tmp in one of the PE version helpers to be unassigned in certain
cases, which is bad when the method relies on it being set.