Add tasks for /tmp/localbuilder dir
This commit adds two new tasks for creating (and cleaning up) a
localbuilder directory that lives in /tmp. For now, this directory will
essentially just contain the custom packages that we will use to build
our custom tarballs. These will be used in the upcoming plan for
creating a custom PE build

This commit also updates the various package building tasks to put the
packages they create into the new localbuilder dir