The build_pe plan arrives (sort of)
This commit adds a plan to build a PE tarball based on provided custom
packages. Currently, the plan doesn't actually build any of the packages
(lol but it will eventually), and just relies on the custom packages
existing in the /tmp/localbuilder/packages directory

This commit also adds the task to correctly sign the custom packages on
el platforms (specifically just el-7 at this point, lol again). This
plan will be expanded in the future so that it's not weirdly hardcoded
to be el-7 specific. Right now, this is just a proof-of-concept to show
that the plan can actually sign packages and be successful.