Update plan to actually have parameters; build pe-modules
This commit adds some more functionality to the plan. It (currently) has
its platform and version hardcoded for convenience, but it now provides
all the parameters requires for building the 4 vanagon projects
currently available in the repository (install, modules, tasks,
backup-tools).

The plan can also now build the pe-modules package. The other packages
should be relatively trivial to add now that this has been completed.
The plan code is pretty ugly right now, and the block of parameters is
pretty enormous, but I'm not sure how avoidable that is. The plan
continues to be something of a work-in-progress/proof-of-concept, but
this commit is an improvement