Large refactor of the `build_<package>` tasks
This commit removes all the specific package building tasks, and
replaces them with one generic `build_vanagon_package` task. This wasn't
really possible before because all of the various package tasks needed
their own specific list of parameters. However, now that there's a plan
that just has a really ugly, gigantic list of parameters, I can pass the
required parameters in as hashes, reducing the `build_vanagon_package`
task's list of parameters down to just a few. This should be a pretty
good improvement, as it removes a whole lot of near-duplicate code.

This commit doesn't update the plan to use this task yet (lol) so the
plan is broken as of this commit, but the plan update will come in the
next commit.