Update build_pe plan and fix `check_parameter` task
This commit updates the build_pe plan to use the new `build_vanagon
package` task, and also adds the ability to build the pe-installer
package as well. It should still be trivial to add pe-tasks package
building and pe-backup-tools package building to the plan, but I'm being
lazy at the moment. el-7-x86_64 is still the only platform for which
the build_pe plan works, but it should work for all PE versions

This commit also updates the `check_parameters` task to actually return
the correct result, because it wasn't doing that before :facepalm: