Add pe-tasks, pe-backup-tools packages to build_pe plan
This commit adds the last two vanagon packages currently supported for
the build_pe plan. There's a lot of duplicate code in this plan right
now, and outside of building the parameter hashes for each package, the
`check_parameters` calls and the following if-elsif blocks can probably
be consolidated into a separate plan that gets called from this one,
which I imagine would reduce the amount of code duplication currently
happening here. Ideally, I'll be doing that in the next commit (or, you
know, relatively soon^TM)