Make multiple PR functionality actually work
This commit ensures that specifying multiple comma-delimited PRs for
some <repo>_pr parameter will actually work. I'm sure this is a bit of a
gross way to do this, and also if there are merge conflicts I'm sure it
will explode spectacularly. For now, though, this does seem to work, so
I think it's alright.