Fix bug that caused single PRs to not be handled
This commit fixes an issue caused by updating the plan to be able to
handle multiple PRs being passed in for a single component. The issue
was that those changes broke functionality for when just _one_ PR was
passed in. This commit adds some handling to ensure that users can still
just pass in a comma-delimited list (so much easier than a real array
from the command line from a user perspective, I swear), or a single PR
and it'll just work:TM: