Add task for signing packages on Debian
This commit adds a working task to sign packages on Debian. While it has
not currently been incorporated into the build_pe plan for real, it's
been tested and verified, and it works as expected. So theoretically
coming soon it will be possible to use this plan to build on master
platforms beyond el-7