Add actual support for building on el-6+ and ubuntu
This commit adds a new plan for determining which package signing method
to use (based on the platform, of course) and now can handle signing
packages on both el and ubuntu (the functionality for this already
existed in the tasks but wasn't actually incorporated into the plan
yet). A known issue is that system Ruby on el-6 is too old to run these
tasks, so while theoretically signing would work if that issue didn't
exist, the issue exists so it's currently broken.