Only run pe-installer-shim plan when necessary
I neglected to add logic to not run the pe-installer-shim build plan
when there aren't any pe-installer-shim changes, so `build_pe` would fail
because it would look for changes that weren't given. This commit fixes
that, so now it'll only run when it should.

This commit also cleans up an accidental file that I didn't mean to
commit :facepalm: